@@ -4,15 +4,18 @@ public class Edge {
 
     private int weight;
 
-    private String name;
-    private Node origin;
-    private Node destination;
+    private final String name;
+    private final Node origin;
+    private final Node destination;
 
-   public Edge(String name, Node origin, Node destination, int weight) {
+    private final String travelType;
+
+    public Edge(String name, Node origin, Node destination, int weight, String travelType) {
         this.name = name;
         this.origin = origin;
         this.destination = destination;
         this.weight = weight;
+        this.travelType = travelType;
     }
 
     public Node getDestination() {
@@ -27,17 +30,16 @@ public class Edge {
         weight = newWeight;
     }
 
+    public String getTravelType() {
+        return travelType;
+    }
+
     public String getName() {
         return name;
     }
 
     @Override
     public String toString() {
-        return "Edge{" +
-                "name='" + name + '\'' +
-                ", origin=" + origin +
-                ", destination=" + destination +
-                ", weight=" + weight +
-                '}';
+        return "%s to %s;\n".formatted(travelType, destination.getName());
     }
 }
